xrCompress: now build on Linux

+++ src/xrCore/crc32.cpp
@@ -48,7 +48,7 @@ private:
     }
 };
 
-u32 crc32(const void* P, u32 len)
+u32 crc32(const void* P, size_t len)
 {
     Crc32Initializer::init();
 
@@ -74,7 +74,7 @@ u32 crc32(const void* P, u32 len)
     return ulCRC ^ 0xffffffff;
 }
 
-u32 crc32(const void* P, u32 len, u32 starting_crc)
+u32 crc32(const void* P, size_t len, u32 starting_crc)
 {
     Crc32Initializer::init();
 
@@ -87,7 +87,7 @@ u32 crc32(const void* P, u32 len, u32 starting_crc)
     return ulCRC ^ 0xffffffff;
 }
 
-u32 path_crc32(const char* path, u32 len)
+u32 path_crc32(const char* path, size_t len)
 {
     Crc32Initializer::init();
